Refactor Preprocess class to include spacy language model and enable preprocessing
